Mon, Feb 7, 2022, 11:31 PM +00:00

+++ filter-less/helpers.c
@@ -4,7 +4,7 @@
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
-    int average;
+    float average;
     for (int i = 0; i < height - 1; i++)
     {
         for (int j = 0; j < width - 1; j++)
